+++ packages/visual-grid-client/src/sdk/makeClose.js
@@ -28,7 +28,7 @@ function makeClose({
       return settleError([])
     }
 
-    const batchId = wrappers[0].getUserSetBatchId()
+    const batchId = wrappers[0].getBatchIdWithoutGenerating()
     globalState.batchStore.addId(batchId)
 
     return waitAndResolveTests(async (testIndex, checkWindowResult) => {