+++ packages/eyes-images/CHANGELOG.md
@@ -5,6 +5,7 @@
 - remove yarn workspaces
 - deprecate `saveDebugData`
 - updated to @applitools/eyes-sdk-core@12.2.6 (from 11.3.4)
+- updated to @applitools/eyes-sdk-core@12.5.0 (from 12.2.6)
 
 
 ## 4.14.2 - 2020/7/8