+++ packages/eyes-images/CHANGELOG.md
@@ -2,10 +2,16 @@
 
 ## Unreleased
 
+
+
+
+## 4.15.0 - 2020/10/22
+
+- adding ability to close batches using batch IDs
 - remove yarn workspaces
 - deprecate `saveDebugData`
 - updated to @applitools/eyes-sdk-core@12.2.6 (from 11.3.4)
-
+- updated to @applitools/eyes-sdk-core@12.5.0 (from 12.2.6)
 
 ## 4.14.2 - 2020/7/8